fix(printf): now ft_printf passes all tests in PrintfTester

+++ libft/includes/libft.h
@@ -91,7 +91,7 @@ char	*ft_get_next_line(int fd);
 size_t	ft_lputaddress_fd(size_t nbr, int fd);
 size_t	ft_lputbin_fd(size_t nbr, int fd);
 size_t	ft_lputdouble_fd(double nbr, int fd);
-size_t	ft_lputhexa_fd(unsigned long long nbr, int fd, char specifier);
+size_t	ft_lputhexa_fd(size_t nbr, int fd, char specifier);
 size_t	ft_lputnbr_fd(long nbr, int fd);
 size_t	ft_lputoctal_fd(size_t nbr, int fd);
 size_t	ft_lputstr_fd(char *s, int fd);
